1.06

+++ Flat.cs
@@ -14,7 +14,7 @@ namespace RealEstateProject
         public string FlatStandard { get; set; }
         public override int RealEstateID { get => base.RealEstateID; set => base.RealEstateID = value; }
 
-        public Flat(int floorNumber, int roomsNumber, FlatStandards flatStandard, double price, double surface, Cities city, double rent, Markets market)
+        public Flat(int floorNumber, int roomsNumber, FlatStandards flatStandard, double price, float surface, Cities city, double rent, Markets market)
             : base(price, surface, city, rent, market)
         {
             this.FloorNumber = floorNumber;
