+++ Plot.cs
@@ -12,7 +12,7 @@ namespace RealEstateProject
         public string RealEstateType { get; set; }
         public override int RealEstateID { get => base.RealEstateID; set => base.RealEstateID = value; }
 
-        public Plot(PlotTypes plotTypeID, double price, double surface, Cities city)
+        public Plot(PlotTypes plotTypeID, double price, float surface, Cities city)
             : base(price, surface, city)
         {
             this.RealEstateID = RealEstate.RealEstateCount;
